Подсчёт количества проданных билетов.

+++ src/Entity/Ticket.php
@@ -33,7 +33,7 @@ class Ticket
     private ?string $place = null;
 
     #[ORM\Column]
-    private ?float $price = null;
+    private ?int $price = null;
 
     #[ORM\JoinColumn(nullable: true, onDelete: 'SET NULL')]
     #[ORM\ManyToOne(inversedBy: 'tickets')]
@@ -68,12 +68,12 @@ class Ticket
         return $this;
     }
 
-    public function getPrice(): ?float
+    public function getPrice(): ?int
     {
         return $this->price;
     }
 
-    public function setPrice(float $price): static
+    public function setPrice(int $price): static
     {
         $this->price = $price;
